# Conflicts:
#   frontend/app/src/main/java/com/ssafy/s309/ui/screen/main/MyPageScreen.kt

+++ frontend/app/src/main/java/com/ssafy/s309/data/repository/source/SamsungHealthDataSource.kt
@@ -36,7 +36,7 @@ class SamsungHealthDataSource
         override suspend fun getTodaySummary(): DailyHealthSummary? {
             val mgr = holder.manager ?: return null
             return try {
-                val steps = mgr.getTodaySteps()
+                val steps = mgr.getTodaySteps().toInt()
                 val sleep = mgr.getLastSleepDurationMinutes()
                 // 권한 미부여 / 데이터 없음 시 둘 다 0 → 데이터 없음으로 간주 → mock fallback
                 if (steps == 0L && sleep == 0) {
